fix: Avoid heavy imports in operator constructor
After a previous commit [1] improving import time on operator import,
this commit fixes the import time issues when instantiating any of the
dbt operators.

The main issue still present was the enum imports in the operator
constructors, which make DAGs slow to import any time they instantiated
one of them.

Profiling can be easily run locally, with the following command:
```shell
python -X importtime -c "from airflow_dbt_python.operators.dbt import DbtRunOperator; DbtRunOperator(task_id='test')" 2>import-times.log
```

[1] https://github.com/tomasfarias/airflow-dbt-python/commit/67c8d7866235e105675bd208659aac2ccc26070e
